extracted serialization helper methods
